requirements.txt
Django==4.1.7
psycopg2-binary==2.9.5
crispy-bootstrap5==0.7
Pillow==9.4.0
djangorestframework==3.14.0

pytest==7.2.2
pytest-django>=4.5.2



